fix: check if song has cover then remove it

+++ app/Http/Controllers/api/v1/SongController.php
@@ -103,7 +103,9 @@ class SongController extends Controller
             $song->delete();
 
             Storage::disk('public')->delete($songPath);
-            Storage::disk('public')->delete($coverPath);
+
+            if ($coverPath)
+                Storage::disk('public')->delete($coverPath);
 
             return response()->json([
                 'message' => 'Song deleted successfully',
